requirements.txt
# Phigros Query 插件依赖
# 安装命令: pip install -r requirements.txt

# 核心依赖
aiohttp>=3.8.0
Pillow>=10.0.0
qrcode>=7.4.2
PyYAML>=6.0
aiofiles>=23.0.0
cryptography>=40.0.0

# 可选依赖（用于高级渲染模式）
# playwright>=1.40.0  # 如需使用 playwright 渲染模式，请取消注释并运行: playwright install chromium
